reverted split test projects by durable version
