--- test/AnalyzerTests/DurableV1/BindingV1/ClientAnalyzerTests.cs
@@ -1,186 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the MIT License. See LICENSE in the project root for license information.
-
-using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CodeFixes;
-using Microsoft.CodeAnalysis.Diagnostics;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using TestHelper;
-
-namespace Microsoft.Azure.WebJobs.Extensions.DurableTask.Analyzers.Test.Binding
-{
-    [TestClass]
-    public class ClientAnalyzerTests : CodeFixVerifier
-    {
-        private static readonly string DiagnosticId = ClientAnalyzer.DiagnosticId;
-        private static readonly DiagnosticSeverity Severity = ClientAnalyzer.Severity;
-
-        private const string V1ExpectedFix = @"
-using Microsoft.Azure.WebJobs;
-using Microsoft.Extensions.Logging;
-
-namespace ExternalInteraction
-{
-    public static class HireEmployee
-    {
-        [FunctionName(""HireEmployee"")]
-        public static async Task<Application> RunOrchestrator(
-            [OrchestrationClient] DurableOrchestrationClient client,
-            ILogger log)
-            {
-            }
-}";
-
-        [TestMethod]
-        public void DurableClient_V1_NonIssue()
-        {
-            VerifyCSharpDiagnostic(V1ExpectedFix);
-        }
-
-        [TestMethod]
-        public void OrchestrationClient_V1_Object()
-        {
-            var test = @"
-using Microsoft.Azure.WebJobs;
-using Microsoft.Extensions.Logging;
-
-namespace ExternalInteraction
-{
-    public static class HireEmployee
-    {
-        [FunctionName(""HireEmployee"")]
-        public static async Task<Application> RunOrchestrator(
-            [OrchestrationClient] Object client,
-            ILogger log)
-            {
-            }
-}";
-            var expectedDiagnostics = new DiagnosticResult
-            {
-                Id = DiagnosticId,
-                Message = string.Format(Resources.V1ClientAnalyzerMessageFormat, "Object"),
-                Severity = Severity,
-                Locations =
-                 new[] {
-                            new DiagnosticResultLocation("Test0.cs", 11, 35)
-                     }
-            };
-
-            VerifyCSharpDiagnostic(test, expectedDiagnostics);
-
-            VerifyCSharpFix(test, V1ExpectedFix, allowNewCompilerDiagnostics: true);
-        }
-
-        [TestMethod]
-        public void OrchestrationClient_V1_String()
-        {
-            var test = @"
-using Microsoft.Azure.WebJobs;
-using Microsoft.Extensions.Logging;
-
-namespace ExternalInteraction
-{
-    public static class HireEmployee
-    {
-        [FunctionName(""HireEmployee"")]
-        public static async Task<Application> RunOrchestrator(
-            [OrchestrationClient] string client,
-            ILogger log)
-            {
-            }
-}";
-            var expectedDiagnostics = new DiagnosticResult
-            {
-                Id = DiagnosticId,
-                Message = string.Format(Resources.V1ClientAnalyzerMessageFormat, "string"),
-                Severity = Severity,
-                Locations =
-                 new[] {
-                            new DiagnosticResultLocation("Test0.cs", 11, 35)
-                     }
-            };
-
-            VerifyCSharpDiagnostic(test, expectedDiagnostics);
-            
-            VerifyCSharpFix(test, V1ExpectedFix, allowNewCompilerDiagnostics: true);
-        }
-
-        [TestMethod]
-        public void OrchestrationClient_V1_V2DurableInterface()
-        {
-            var test = @"
-using Microsoft.Azure.WebJobs;
-using Microsoft.Extensions.Logging;
-
-namespace ExternalInteraction
-{
-    public static class HireEmployee
-    {
-        [FunctionName(""HireEmployee"")]
-        public static async Task<Application> RunOrchestrator(
-            [OrchestrationClient] IDurableClient client,
-            ILogger log)
-            {
-            }
-}";
-            var expectedDiagnostics = new DiagnosticResult
-            {
-                Id = DiagnosticId,
-                Message = string.Format(Resources.V1ClientAnalyzerMessageFormat, "IDurableClient"),
-                Severity = Severity,
-                Locations =
-                 new[] {
-                            new DiagnosticResultLocation("Test0.cs", 11, 35)
-                     }
-            };
-
-            VerifyCSharpDiagnostic(test, expectedDiagnostics);
-
-            VerifyCSharpFix(test, V1ExpectedFix, allowNewCompilerDiagnostics: true);
-        }
-
-        [TestMethod]
-        public void OrchestrationClient_V1_Tuple()
-        {
-            var test = @"
-using Microsoft.Azure.WebJobs;
-using Microsoft.Extensions.Logging;
-
-namespace ExternalInteraction
-{
-    public static class HireEmployee
-    {
-        [FunctionName(""HireEmployee"")]
-        public static async Task<Application> RunOrchestrator(
-            [OrchestrationClient] Tuple<int, string> client,
-            ILogger log)
-            {
-            }
-}";
-            var expectedDiagnostics = new DiagnosticResult
-            {
-                Id = DiagnosticId,
-                Message = string.Format(Resources.V1ClientAnalyzerMessageFormat, "Tuple<int, string>"),
-                Severity = Severity,
-                Locations =
-                 new[] {
-                            new DiagnosticResultLocation("Test0.cs", 11, 35)
-                     }
-            };
-
-            VerifyCSharpDiagnostic(test, expectedDiagnostics);
-
-            VerifyCSharpFix(test, V1ExpectedFix, allowNewCompilerDiagnostics: true);
-        }
-
-        protected override CodeFixProvider GetCSharpCodeFixProvider()
-        {
-            return new ClientCodeFixProvider();
-        }
-
-        protected override DiagnosticAnalyzer GetCSharpDiagnosticAnalyzer()
-        {
-            return new ClientAnalyzer();
-        }
-    }
-}
