+++ test/WebJobs.Extensions.DurableTask.Analyzers.Test/Binding/ClientAnalyzerTests.cs
@@ -15,6 +15,22 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask.Analyzers.Test.Binding
         private static readonly string DiagnosticId = ClientAnalyzer.DiagnosticId;
         private static readonly DiagnosticSeverity Severity = ClientAnalyzer.Severity;
 
+        private const string V1ExpectedFix = @"
+using Microsoft.Azure.WebJobs;
+using Microsoft.Extensions.Logging;
+
+namespace ExternalInteraction
+{
+    public static class HireEmployee
+    {
+        [FunctionName(""HireEmployee"")]
+        public static async Task<Application> RunOrchestrator(
+            [OrchestrationClient] DurableOrchestrationClient client,
+            ILogger log)
+            {
+            }
+}";
+
         private const string V2ClientExpectedFix = @"
 using Microsoft.Azure.WebJobs.Extensions.DurableTask;
 using Microsoft.Extensions.Logging;
@@ -64,11 +80,165 @@ namespace ExternalInteraction
 }";
 
         [TestMethod]
+        public void DurableClient_V1_NonIssue()
+        {
+            SyntaxNodeUtils.version = DurableVersion.V1;
+
+            VerifyCSharpDiagnostic(V1ExpectedFix);
+        }
+
+        [TestMethod]
         public void DurableClient_V2_NonIssue()
         {
+            SyntaxNodeUtils.version = DurableVersion.V2;
+
             VerifyCSharpDiagnostic(V2ClientExpectedFix);
             VerifyCSharpDiagnostic(V2OrchestrationClientExpectedFix);
             VerifyCSharpDiagnostic(V2EntityClientExpectedFix);
+        }
+
+        [TestMethod]
+        public void OrchestrationClient_V1_Object()
+        {
+            var test = @"
+using Microsoft.Azure.WebJobs;
+using Microsoft.Extensions.Logging;
+
+namespace ExternalInteraction
+{
+    public static class HireEmployee
+    {
+        [FunctionName(""HireEmployee"")]
+        public static async Task<Application> RunOrchestrator(
+            [OrchestrationClient] Object client,
+            ILogger log)
+            {
+            }
+}";
+            var expectedDiagnostics = new DiagnosticResult
+            {
+                Id = DiagnosticId,
+                Message = string.Format(Resources.V1ClientAnalyzerMessageFormat, "Object"),
+                Severity = Severity,
+                Locations =
+                 new[] {
+                            new DiagnosticResultLocation("Test0.cs", 11, 35)
+                     }
+            };
+
+            SyntaxNodeUtils.version = DurableVersion.V1;
+
+            VerifyCSharpDiagnostic(test, expectedDiagnostics);
+
+            VerifyCSharpFix(test, V1ExpectedFix);
+        }
+
+        [TestMethod]
+        public void OrchestrationClient_V1_String()
+        {
+            var test = @"
+using Microsoft.Azure.WebJobs;
+using Microsoft.Extensions.Logging;
+
+namespace ExternalInteraction
+{
+    public static class HireEmployee
+    {
+        [FunctionName(""HireEmployee"")]
+        public static async Task<Application> RunOrchestrator(
+            [OrchestrationClient] string client,
+            ILogger log)
+            {
+            }
+}";
+            var expectedDiagnostics = new DiagnosticResult
+            {
+                Id = DiagnosticId,
+                Message = string.Format(Resources.V1ClientAnalyzerMessageFormat, "string"),
+                Severity = Severity,
+                Locations =
+                 new[] {
+                            new DiagnosticResultLocation("Test0.cs", 11, 35)
+                     }
+            };
+
+            SyntaxNodeUtils.version = DurableVersion.V1;
+
+            VerifyCSharpDiagnostic(test, expectedDiagnostics);
+            
+            VerifyCSharpFix(test, V1ExpectedFix, allowNewCompilerDiagnostics: true);
+        }
+
+        [TestMethod]
+        public void OrchestrationClient_V1_V2DurableInterface()
+        {
+            var test = @"
+using Microsoft.Azure.WebJobs;
+using Microsoft.Extensions.Logging;
+
+namespace ExternalInteraction
+{
+    public static class HireEmployee
+    {
+        [FunctionName(""HireEmployee"")]
+        public static async Task<Application> RunOrchestrator(
+            [OrchestrationClient] IDurableClient client,
+            ILogger log)
+            {
+            }
+}";
+            var expectedDiagnostics = new DiagnosticResult
+            {
+                Id = DiagnosticId,
+                Message = string.Format(Resources.V1ClientAnalyzerMessageFormat, "IDurableClient"),
+                Severity = Severity,
+                Locations =
+                 new[] {
+                            new DiagnosticResultLocation("Test0.cs", 11, 35)
+                     }
+            };
+
+            SyntaxNodeUtils.version = DurableVersion.V1;
+
+            VerifyCSharpDiagnostic(test, expectedDiagnostics);
+
+            VerifyCSharpFix(test, V1ExpectedFix);
+        }
+
+        [TestMethod]
+        public void OrchestrationClient_V1_Tuple()
+        {
+            var test = @"
+using Microsoft.Azure.WebJobs;
+using Microsoft.Extensions.Logging;
+
+namespace ExternalInteraction
+{
+    public static class HireEmployee
+    {
+        [FunctionName(""HireEmployee"")]
+        public static async Task<Application> RunOrchestrator(
+            [OrchestrationClient] Tuple<int, string> client,
+            ILogger log)
+            {
+            }
+}";
+            var expectedDiagnostics = new DiagnosticResult
+            {
+                Id = DiagnosticId,
+                Message = string.Format(Resources.V1ClientAnalyzerMessageFormat, "Tuple<int, string>"),
+                Severity = Severity,
+                Locations =
+                 new[] {
+                            new DiagnosticResultLocation("Test0.cs", 11, 35)
+                     }
+            };
+
+            SyntaxNodeUtils.version = DurableVersion.V1;
+
+            VerifyCSharpDiagnostic(test, expectedDiagnostics);
+
+            VerifyCSharpFix(test, V1ExpectedFix);
         }
 
         [TestMethod]
@@ -100,11 +270,13 @@ namespace ExternalInteraction
                      }
             };
 
+            SyntaxNodeUtils.version = DurableVersion.V2;
+
             VerifyCSharpDiagnostic(test, expectedDiagnostics);
 
-            VerifyCSharpFix(test, V2ClientExpectedFix, 0, allowNewCompilerDiagnostics: true);
-            VerifyCSharpFix(test, V2EntityClientExpectedFix, 1, allowNewCompilerDiagnostics: true);
-            VerifyCSharpFix(test, V2OrchestrationClientExpectedFix, 2, allowNewCompilerDiagnostics: true);
+            VerifyCSharpFix(test, V2ClientExpectedFix, 0);
+            VerifyCSharpFix(test, V2EntityClientExpectedFix, 1);
+            VerifyCSharpFix(test, V2OrchestrationClientExpectedFix, 2);
         }
 
         [TestMethod]
@@ -136,11 +308,13 @@ namespace ExternalInteraction
                      }
             };
 
+            SyntaxNodeUtils.version = DurableVersion.V2;
+
             VerifyCSharpDiagnostic(test, expectedDiagnostics);
 
-            VerifyCSharpFix(test, V2ClientExpectedFix, 0, allowNewCompilerDiagnostics: true);
-            VerifyCSharpFix(test, V2EntityClientExpectedFix, 1, allowNewCompilerDiagnostics: true);
-            VerifyCSharpFix(test, V2OrchestrationClientExpectedFix, 2, allowNewCompilerDiagnostics: true);
+            VerifyCSharpFix(test, V2ClientExpectedFix, 0);
+            VerifyCSharpFix(test, V2EntityClientExpectedFix, 1);
+            VerifyCSharpFix(test, V2OrchestrationClientExpectedFix, 2);
         }
 
         [TestMethod]
@@ -172,11 +346,13 @@ namespace ExternalInteraction
                      }
             };
 
+            SyntaxNodeUtils.version = DurableVersion.V2;
+
             VerifyCSharpDiagnostic(test, expectedDiagnostics);
 
-            VerifyCSharpFix(test, V2ClientExpectedFix, 0, allowNewCompilerDiagnostics: true);
-            VerifyCSharpFix(test, V2EntityClientExpectedFix, 1, allowNewCompilerDiagnostics: true);
-            VerifyCSharpFix(test, V2OrchestrationClientExpectedFix, 2, allowNewCompilerDiagnostics: true);
+            VerifyCSharpFix(test, V2ClientExpectedFix, 0);
+            VerifyCSharpFix(test, V2EntityClientExpectedFix, 1);
+            VerifyCSharpFix(test, V2OrchestrationClientExpectedFix, 2);
         }
 
         [TestMethod]
@@ -208,11 +384,13 @@ namespace ExternalInteraction
                      }
             };
 
+            SyntaxNodeUtils.version = DurableVersion.V2;
+
             VerifyCSharpDiagnostic(test, expectedDiagnostics);
 
-            VerifyCSharpFix(test, V2ClientExpectedFix, 0, allowNewCompilerDiagnostics: true);
-            VerifyCSharpFix(test, V2EntityClientExpectedFix, 1, allowNewCompilerDiagnostics: true);
-            VerifyCSharpFix(test, V2OrchestrationClientExpectedFix, 2, allowNewCompilerDiagnostics: true);
+            VerifyCSharpFix(test, V2ClientExpectedFix, 0);
+            VerifyCSharpFix(test, V2EntityClientExpectedFix, 1);
+            VerifyCSharpFix(test, V2OrchestrationClientExpectedFix, 2);
         }
 
         protected override CodeFixProvider GetCSharpCodeFixProvider()
