+++ test/WebJobs.Extensions.DurableTask.Analyzers.Test/Binding/EntityContentAnalyzerTests.cs
