+++ test/WebJobs.Extensions.DurableTask.Analyzers.Test/Entity/ClassNameAnalyzerTests.cs
