+++ test/WebJobs.Extensions.DurableTask.Analyzers.Test/Entity/DispatchEntityNameAnalyzerTests.cs
