+++ test/WebJobs.Extensions.DurableTask.Analyzers.Test/Entity/StaticFunctionAnalyzerTests.cs
