+++ test/WebJobs.Extensions.DurableTask.Analyzers.Test/EntityInterface/EntityInterfaceReturnTypeAnalyzerTests.cs
