+++ test/WebJobs.Extensions.DurableTask.Analyzers.Test/EntityInterface/InterfaceAnalyzerTests.cs
