+++ test/WebJobs.Extensions.DurableTask.Analyzers.Test/Helpers/DiagnosticResult.cs
